pyproject.toml
[project]
name = "seguridad-ciudadana-unasam"
version = "2.0.0"
description = "Sistema Inteligente de Seguridad Ciudadana - Predicción de Criminalidad con ML"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # ── ML & Data ──
    "pandas>=2.0",
    "pyarrow>=14.0",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "numpy>=1.26",

    # ── API ──
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # ── Frontend ──
    "streamlit>=1.30",
    "plotly>=5.18",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-asyncio>=0.23",
]
ingestion = [
    "google-cloud-bigquery>=3.0",
    "google-cloud-bigquery-storage>=2.0",
    "pandas-gbq>=0.22",
    "db-dtypes>=1.2",
]
analysis = [
    "seaborn>=0.13",
    "matplotlib>=3.8",
    "scipy>=1.12",
    "statsmodels>=0.14",
    "pmdarima>=2.0",
    "xgboost>=2.0",
    "tabulate>=0.9",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["backend*", "frontend*"]
